changed health and projectile
health now uses c# events instead of unity events. Projectile now is destroyed after hiting a wall and is much faster
